refactor: entity mutation ws subscription hook
